Student Management System
Here is the first commit, it was to see if we could make a log in page for the user, admin, as well as the initial database, that includes the admin's log in information, as well as the user's log in information.

From the admin side, it was also tested to see if menus to add students and the sort could be added.